# add sync.Pool to get hash.Hash where possible