--- a/algo_hs.go
+++ b/algo_hs.go
@@ -3,6 +3,7 @@ package jwt
 import (
 	"crypto"
 	"crypto/hmac"
+	"hash"
 )
 
 // NewSignerHS returns a new HMAC-based signer.
@@ -15,9 +16,10 @@ func NewSignerHS(alg Algorithm, key []byte) (Signer, error) {
 		return nil, err
 	}
 	return &hsAlg{
-		alg:  alg,
-		hash: hash,
-		key:  key,
+		alg:      alg,
+		hash:     hash,
+		key:      key,
+		hashPool: newHashPool(),
 	}, nil
 }
 
@@ -31,9 +33,10 @@ func NewVerifierHS(alg Algorithm, key []byte) (Verifier, error) {
 		return nil, err
 	}
 	return &hsAlg{
-		alg:  alg,
-		hash: hash,
-		key:  key,
+		alg:      alg,
+		hash:     hash,
+		key:      key,
+		hashPool: newHashPool(),
 	}, nil
 }
 
@@ -51,9 +54,10 @@ func getHashHMAC(alg Algorithm) (crypto.Hash, error) {
 }
 
 type hsAlg struct {
-	alg  Algorithm
-	hash crypto.Hash
-	key  []byte
+	alg      Algorithm
+	hash     crypto.Hash
+	key      []byte
+	hashPool *hashPool
 }
 
 func (h hsAlg) Algorithm() Algorithm {
@@ -76,7 +80,10 @@ func (h hsAlg) Verify(payload, signature []byte) error {
 }
 
 func (h hsAlg) sign(payload []byte) ([]byte, error) {
-	hasher := hmac.New(h.hash.New, h.key)
+	hasher := h.hashPool.getHash(func() hash.Hash {
+		return hmac.New(h.hash.New, h.key)
+	})
+	defer h.hashPool.putHash(hasher)
 
 	_, err := hasher.Write(payload)
 	if err != nil {