@storozhukBM 2020-05-08
@FZambia Looks good.
I'd suggest you run all benchmarks with `-count=10` then `benchstat` will show us the variance of the measurements. Then I'd use it only for algorithms that show some measurable difference. Maybe it worth removing the pool for EdDSA and maybe RS family, but it will be more evident after remeasurements.