@FZambia 2020-05-09
Yeah - it's definetely worth removing for EdDSA and RSA since most time and mem spent inside other calls.