@FZambia 2020-05-09
Interesting, I was expecting at least notable improvement in algorithms other then HMAC, but actually it only makes sense for HMAC. Here is benchstat after 10 tests:
```
name                 old time/op    new time/op    delta
EDSA/Sign-EdDSA-4      65.8µs ± 1%    65.8µs ± 1%     ~     (p=0.780 n=9+10)
EDSA/Verify-EdDSA-4     167µs ± 3%     168µs ± 3%     ~     (p=0.529 n=10+10)
ES/Sign-ES256-4        35.1µs ± 2%    35.1µs ± 1%     ~     (p=0.829 n=10+8)
ES/Verify-ES256-4      97.8µs ± 4%   105.4µs ± 5%   +7.80%  (p=0.000 n=10+10)
ES/Sign-ES384-4        5.22ms ± 1%    5.50ms ± 6%   +5.28%  (p=0.000 n=9+10)
ES/Verify-ES384-4      11.4ms ± 4%    11.4ms ± 3%     ~     (p=0.971 n=10+10)
ES/Sign-ES512-4        9.34ms ± 3%    9.83ms ±16%     ~     (p=0.393 n=10+10)
ES/Verify-ES512-4      22.0ms ± 6%    20.9ms ±18%     ~     (p=0.549 n=9+10)
PS/Sign-PS256-4        1.83ms ± 1%    1.87ms ± 1%   +2.22%  (p=0.000 n=10+10)
PS/Verify-PS256-4      83.5µs ± 3%    83.7µs ± 1%     ~     (p=0.315 n=10+9)
PS/Sign-PS384-4        1.83ms ± 1%    1.87ms ± 3%   +2.36%  (p=0.000 n=10+10)
PS/Verify-PS384-4      81.4µs ± 1%    83.2µs ± 1%   +2.26%  (p=0.000 n=9+10)
PS/Sign-PS512-4        1.84ms ± 2%    1.87ms ± 2%   +1.19%  (p=0.008 n=10+9)
PS/Verify-PS512-4      82.5µs ± 2%    80.4µs ± 1%   -2.54%  (p=0.000 n=9+9)
RS/Sign-RS256-4        1.86ms ± 1%    1.83ms ± 1%   -1.50%  (p=0.001 n=8+9)
RS/Verify-RS256-4      80.1µs ± 1%    80.4µs ± 0%   +0.47%  (p=0.015 n=9+8)
RS/Sign-RS384-4        1.87ms ± 3%    1.85ms ± 1%     ~     (p=0.315 n=10+9)
RS/Verify-RS384-4      79.6µs ± 2%    77.6µs ± 0%   -2.46%  (p=0.000 n=9+9)
RS/Sign-RS512-4        1.85ms ± 1%    1.84ms ± 1%   -0.68%  (p=0.017 n=10+9)
RS/Verify-RS512-4      80.4µs ± 2%    77.7µs ± 1%   -3.33%  (p=0.000 n=10+9)
HS/Sign-HS256-4        3.24µs ± 3%    2.85µs ± 1%  -11.89%  (p=0.000 n=9+9)
HS/Verify-HS256-4      1.62µs ± 1%    1.29µs ± 1%  -19.90%  (p=0.000 n=9+10)
HS/Sign-HS384-4        3.54µs ± 1%    3.00µs ± 1%  -15.15%  (p=0.000 n=9+9)
HS/Verify-HS384-4      1.90µs ± 1%    1.40µs ± 3%  -26.09%  (p=0.000 n=10+9)
HS/Sign-HS512-4        3.58µs ± 1%    3.05µs ± 1%  -14.89%  (p=0.000 n=10+10)
HS/Verify-HS512-4      1.91µs ± 1%    1.42µs ± 1%  -25.73%  (p=0.000 n=9+9)

name                 old alloc/op   new alloc/op   delta
EDSA/Sign-EdDSA-4      1.42kB ± 0%    1.42kB ± 0%     ~     (all equal)
EDSA/Verify-EdDSA-4      294B ± 0%      294B ± 0%     ~     (p=0.081 n=10+10)
ES/Sign-ES256-4        3.95kB ± 0%    3.83kB ± 0%   -3.23%  (p=0.000 n=10+10)
ES/Verify-ES256-4      1.26kB ± 0%    1.13kB ± 0%  -10.12%  (p=0.000 n=10+10)
ES/Sign-ES384-4        1.75MB ± 0%    1.75MB ± 0%     ~     (p=0.393 n=10+10)
ES/Verify-ES384-4      3.86MB ± 4%    3.86MB ± 3%     ~     (p=0.912 n=10+10)
ES/Sign-ES512-4        3.03MB ± 0%    3.03MB ± 0%     ~     (p=0.796 n=10+10)
ES/Verify-ES512-4      7.33MB ± 6%    6.86MB ±18%     ~     (p=0.447 n=9+10)
PS/Sign-PS256-4        32.9kB ± 0%    32.8kB ± 0%   -0.37%  (p=0.000 n=10+10)
PS/Verify-PS256-4      5.89kB ± 0%    5.76kB ± 0%   -2.15%  (p=0.000 n=10+10)
PS/Sign-PS384-4        33.1kB ± 0%    32.9kB ± 0%   -0.66%  (p=0.000 n=10+10)
PS/Verify-PS384-4      6.13kB ± 0%    5.91kB ± 0%   -3.61%  (p=0.000 n=10+10)
PS/Sign-PS512-4        33.1kB ± 0%    32.9kB ± 0%   -0.65%  (p=0.000 n=10+10)
PS/Verify-PS512-4      6.18kB ± 0%    5.95kB ± 0%   -3.63%  (p=0.000 n=10+9)
RS/Sign-RS256-4        32.2kB ± 0%    32.1kB ± 0%   -0.39%  (p=0.000 n=10+10)
RS/Verify-RS256-4      5.42kB ± 0%    5.30kB ± 0%   -2.36%  (p=0.000 n=10+10)
RS/Sign-RS384-4        32.4kB ± 0%    32.1kB ± 0%   -0.67%  (p=0.000 n=10+10)
RS/Verify-RS384-4      5.54kB ± 0%    5.31kB ± 0%   -4.06%  (p=0.000 n=10+10)
RS/Sign-RS512-4        32.4kB ± 0%    32.2kB ± 0%   -0.68%  (p=0.000 n=10+10)
RS/Verify-RS512-4      5.55kB ± 0%    5.33kB ± 0%   -4.04%  (p=0.000 n=10+10)
HS/Sign-HS256-4        1.33kB ± 0%    0.85kB ± 0%  -36.14%  (p=0.000 n=10+10)
HS/Verify-HS256-4        512B ± 0%       32B ± 0%  -93.75%  (p=0.000 n=10+10)
HS/Sign-HS384-4        1.70kB ± 0%    0.90kB ± 0%  -47.17%  (p=0.000 n=10+10)
HS/Verify-HS384-4        848B ± 0%       48B ± 0%  -94.34%  (p=0.000 n=10+10)
HS/Sign-HS512-4        1.78kB ± 0%    0.98kB ± 0%  -45.05%  (p=0.000 n=10+10)
HS/Verify-HS512-4        864B ± 0%       64B ± 0%  -92.59%  (p=0.000 n=10+10)

name                 old allocs/op  new allocs/op  delta
EDSA/Sign-EdDSA-4        18.0 ± 0%      18.0 ± 0%     ~     (all equal)
EDSA/Verify-EdDSA-4      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
ES/Sign-ES256-4          51.0 ± 0%      50.0 ± 0%   -1.96%  (p=0.000 n=10+10)
ES/Verify-ES256-4        21.0 ± 0%      20.0 ± 0%   -4.76%  (p=0.000 n=10+10)
ES/Sign-ES384-4         14.4k ± 0%     14.4k ± 0%     ~     (p=0.448 n=10+10)
ES/Verify-ES384-4       31.9k ± 4%     31.9k ± 3%     ~     (p=0.896 n=10+10)
ES/Sign-ES512-4         19.6k ± 0%     19.6k ± 0%     ~     (p=0.699 n=10+10)
ES/Verify-ES512-4       47.4k ± 6%     44.4k ±18%     ~     (p=0.447 n=9+10)
PS/Sign-PS256-4           129 ± 0%       128 ± 0%   -0.78%  (p=0.000 n=10+10)
PS/Verify-PS256-4        19.0 ± 0%      18.0 ± 0%   -5.26%  (p=0.000 n=10+10)
PS/Sign-PS384-4           129 ± 0%       128 ± 0%   -0.78%  (p=0.000 n=10+10)
PS/Verify-PS384-4        19.0 ± 0%      18.0 ± 0%   -5.26%  (p=0.000 n=10+10)
PS/Sign-PS512-4           129 ± 0%       128 ± 0%   -0.78%  (p=0.000 n=10+10)
PS/Verify-PS512-4        19.0 ± 0%      18.0 ± 0%   -5.26%  (p=0.000 n=10+10)
RS/Sign-RS256-4           123 ± 0%       122 ± 0%   -0.81%  (p=0.000 n=10+10)
RS/Verify-RS256-4        13.0 ± 0%      12.0 ± 0%   -7.69%  (p=0.000 n=10+10)
RS/Sign-RS384-4           123 ± 0%       122 ± 0%   -0.81%  (p=0.000 n=10+10)
RS/Verify-RS384-4        13.0 ± 0%      12.0 ± 0%   -7.69%  (p=0.000 n=10+10)
RS/Sign-RS512-4           123 ± 0%       122 ± 0%   -0.81%  (p=0.000 n=10+10)
RS/Verify-RS512-4        13.0 ± 0%      12.0 ± 0%   -7.69%  (p=0.000 n=10+10)
HS/Sign-HS256-4          18.0 ± 0%      13.0 ± 0%  -27.78%  (p=0.000 n=10+10)
HS/Verify-HS256-4        6.00 ± 0%      1.00 ± 0%  -83.33%  (p=0.000 n=10+10)
HS/Sign-HS384-4          18.0 ± 0%      13.0 ± 0%  -27.78%  (p=0.000 n=10+10)
HS/Verify-HS384-4        6.00 ± 0%      1.00 ± 0%  -83.33%  (p=0.000 n=10+10)
HS/Sign-HS512-4          18.0 ± 0%      13.0 ± 0%  -27.78%  (p=0.000 n=10+10)
HS/Verify-HS512-4        6.00 ± 0%      1.00 ± 0%  -83.33%  (p=0.000 n=10+10)
```