@FZambia 2020-05-09
Closing - first measurements in PR were a mistake due to test env inconsistency and I was tricked. Thanks @storozhukBM for pointing at statistics error.